# code/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Funerária Baldan NF API
  version: 0.6.0
  description: |
    API do PDV Baldan (NFC-e).
    F5 — hardening: policies (config admin), user ativo/role, Idempotency-Key em venda, throttle mutations, 403 JSON.
    F4 — módulos: sangria/fechar caixa, consignado, upload A1; UI cadastros/estoque/config.
    F3 — vertical slice: caixa → finalizar venda → job NFC-e (fila fiscal) + XML/DANFE fake.
    F2 — auth Sanctum: login / me / logout / refresh.
    F1 — domínio: marcas, categorias, produtos, clientes/dependentes, config fiscal, movimentação de estoque.
servers:
  - url: http://localhost:8000/api/v1
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
...
paths:
...
  /vendas/finalizar:
    post:
      operationId: finalizarVenda
      tags: [Vendas]
      security: [{ sanctum: [] }]
      summary: Finaliza venda (baixa estoque, cria NFC-e, dispatch job). Header Idempotency-Key evita duplicata.
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string, maxLength: 64 }
      responses:
        '201': { description: Venda criada }
        '200': { description: Replay idempotente (mesma key) }
        '401': { description: Não autenticado }
        '422': { description: Sem caixa / estoque / validação }
...
  /consignados/{consignado}/devolver:
    parameters:
      - in: path
        name: consignado
        required: true
        schema: { type: integer }
    post:
      operationId: devolverConsignado
      tags: [Consignado]
      security: [{ sanctum: [] }]
      summary: Devolve itens e repõe estoque
      responses:
        '200': { description: OK }
        '401': { description: Não autenticado }
        '422': { description: Validação }
...